+++ lib/shared/core/constants/shared_preferences_keys.dart
@@ -1,5 +1,6 @@
 class SharedPreferencesKeys {
   static const String repositoryPath = "repositoryPath";
   static const String repositoryBookmark = "repositoryBookmark";
-  static const diffModeDisplay = 'diff_mode_display';
+  static const String diffModeDisplay = 'diff_mode_display';
+  static const String repositorySidebarWidth = 'repository_sidebar_width';
 }